+++ src/index.ts
@@ -1,7 +1,0 @@
-export { initFederation } from './lib/init-federation';
-export { NFError } from './lib/native-federation.error';
-
-export * from './lib/1.domain';
-export * from './lib/2.app';
-
-export * as config from './lib/4.config';