+++ src/lib/4.config/import-map/use-systemjs.ts
@@ -1,4 +1,4 @@
-import type { ImportMapConfig } from "lib/2.app";
+import type { ImportMapConfig } from "lib/2.app/config/import-map.contract";
 
 
 export const useSystemJSImportMap = (): ImportMapConfig => ({