+++ src/lib/4.config/index.ts
@@ -1,3 +1,4 @@
 export * from "./logging";
 export * from "./storage";
 export * from "./import-map";
+export * from "./mode";