+++ src/lib/4.config/mode/caching.profile.ts
@@ -1,0 +1,6 @@
+
+
+export const cachingProfile = {
+    latestSharedExternal: false,
+    skipCachedRemotes: true
+}